migration_tools_requirements.txt
# Database Migration Tools - Requirements
# Install with: pip install -r migration_tools_requirements.txt

# Core packages
psycopg2-binary>=2.9.5      # PostgreSQL adapter
python-dotenv>=0.21.0       # Environment variable management
sqlalchemy>=1.4.46          # SQL toolkit and ORM
alembic>=1.9.1              # Database migration tool

# Testing and analysis
pytest>=7.2.1               # Testing framework
google-cloud-logging>=3.5.0 # For log analysis script
pandas>=1.5.3               # For data analysis and validation

# Production support
gunicorn>=20.1.0            # WSGI HTTP server
flask-migrate>=4.0.4        # Flask integration with Alembic
pyyaml>=6.0                 # YAML parser for configuration 